+++ sampleapplication/src/main/java/com/byoutline/sampleapplication/draweractivity/DrawerFragments.kt
@@ -7,8 +7,9 @@ import android.view.View
 import android.view.ViewGroup
 import com.byoutline.sampleapplication.R
 import com.byoutline.sampleapplication.databinding.FragmentShowDialogBinding
-import com.byoutline.secretsauce.di.AttachableViewModel
-import com.byoutline.secretsauce.di.lazyViewModelWithAutoLifecycle
+import com.byoutline.secretsauce.databinding.inflateAndSetVM
+import com.byoutline.secretsauce.lifecycle.AttachableViewModel
+import com.byoutline.secretsauce.lifecycle.getVMWithAutoLifecycle
 import java.util.concurrent.atomic.AtomicInteger
 
 class FirstFragment : CountingFragment() {
@@ -20,13 +21,10 @@ class SecondFragment : CountingFragment() {
 }
 
 abstract class CountingFragment : BaseFragment() {
-    val viewModel by lazyViewModelWithAutoLifecycle(CountingViewModel::class)
-    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?,
-                              savedInstanceState: Bundle?): View? {
-        val binding = FragmentShowDialogBinding.inflate(inflater, container, false)
-        binding.viewModel = viewModel
-        return binding.root
-    }
+    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? =
+            inflateAndSetVM<FragmentShowDialogBinding>(inflater, container, R.layout.fragment_show_dialog,
+                    viewModel = getVMWithAutoLifecycle(CountingViewModel::class)
+            ).root
 }
 
 class CountingViewModel : AttachableViewModel<BaseFragment>() {