+++ sampleapplication/src/main/java/com/byoutline/sampleapplication/rx/RxLifecycleActivity.kt
@@ -5,9 +5,9 @@ import android.os.Bundle
 import android.support.v7.app.AppCompatActivity
 import com.byoutline.sampleapplication.R
 import com.byoutline.sampleapplication.databinding.RxlifecycleActivityBinding
-import com.byoutline.secretsauce.di.AttachableViewModel
-import com.byoutline.secretsauce.di.bindContentView
-import com.byoutline.secretsauce.di.lazyViewModelWithAutoLifecycle
+import com.byoutline.secretsauce.databinding.bindContentView
+import com.byoutline.secretsauce.lifecycle.AttachableViewModelRx
+import com.byoutline.secretsauce.lifecycle.getVMWithAutoLifecycle
 import java.util.concurrent.TimeUnit
 import io.reactivex.Observable as RxObservable
 
@@ -17,23 +17,21 @@ import io.reactivex.Observable as RxObservable
  */
 class RxLifecycleActivity : AppCompatActivity() {
 
-    private val viewModel by lazyViewModelWithAutoLifecycle(RxLifecycleViewModel::class)
-
     public override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        val binding = bindContentView<RxlifecycleActivityBinding>(R.layout.rxlifecycle_activity)
-        binding.viewModel = viewModel
+        bindContentView<RxlifecycleActivityBinding>(R.layout.rxlifecycle_activity).apply {
+            viewModel = getVMWithAutoLifecycle(RxLifecycleViewModel::class)
+        }
     }
 }
 
 
-class RxLifecycleViewModel : AttachableViewModel<Any>() {
+class RxLifecycleViewModel : AttachableViewModelRx<Any>() {
     val value = ObservableLong()
     override fun onAttach(view: Any) {
-        val disposable = RxObservable.interval(1, TimeUnit.SECONDS)
+        RxObservable.interval(1, TimeUnit.SECONDS)
                 .subscribe { value.set(value.get() + 1) }
-        super.onAttach(view, onDetachAction = {
-            disposable.dispose()
-        })
+                .disposeOnDetach()
+        super.onAttach(view)
     }
 }